// src/app/models/permission.ts
export enum Permission {
  USER_CREATE = 'USER_CREATE',
  USER_READ = 'USER_READ',
  USER_UPDATE = 'USER_UPDATE',
  USER_DELETE = 'USER_DELETE',

  MACHINE_SEARCH = 'MACHINE_SEARCH',
  MACHINE_START = 'MACHINE_START',
  MACHINE_STOP = 'MACHINE_STOP',
  MACHINE_RESTART = 'MACHINE_RESTART',
  MACHINE_CREATE = 'MACHINE_CREATE',
  MACHINE_DESTROY = 'MACHINE_DESTROY',
}